@@ -2,43 +2,37 @@
     "name": "laravel/laravel",
     "type": "project",
     "description": "The Laravel Framework.",
-    "keywords": ["framework", "laravel"],
+    "keywords": [
+        "framework",
+        "laravel"
+    ],
     "license": "MIT",
     "require": {
         "php": "^8.2",
-        "archtechx/enums": "^0.3.1",
-        "bepsvpt/secure-headers": "^7.2",
-        "caneara/axiom": "^3.2",
-        "doctrine/dbal": "^3.4",
+        "bepsvpt/secure-headers": "^7.4",
+        "doctrine/dbal": "^3.6",
         "guzzlehttp/guzzle": "^7.2",
-        "illuminate/console": "^9.0",
         "itsgoingd/clockwork": "^5.1",
-        "kreait/laravel-firebase": "^5.0",
-        "laravel/framework": "^9.0",
-        "laravel/sanctum": "^3.0",
-        "laravel/slack-notification-channel": "^2.3",
-        "laravel/tinker": "^2.5",
-        "league/flysystem-aws-s3-v3": "^3.3",
-        "psr/log": "~2",
-        "rollbar/rollbar-laravel": "^7.2",
-        "spatie/eloquent-sortable": "^4.0",
-        "spatie/laravel-translatable": "^6.0",
-        "spatie/laravel-translation-loader": "^2.7",
-        "staudenmeir/belongs-to-through": "^2.12"
+        "laravel/framework": "^10.8",
+        "laravel/sanctum": "^3.2",
+        "laravel/slack-notification-channel": "^2.5",
+        "laravel/tinker": "^2.8",
+        "league/flysystem-aws-s3-v3": "^3.15",
+        "rollbar/rollbar-laravel": "^8.0",
+        "spatie/laravel-translatable": "^6.5",
+        "spatie/laravel-translation-loader": "^2.7"
     },
     "require-dev": {
-        "brianium/paratest": "^6.6",
+        "brianium/paratest": "^7.1",
         "fakerphp/faker": "^1.9.1",
-        "laravel-lang/attributes": "^2.0",
-        "laravel-lang/lang": "^12.0",
-        "laravel-lang/publisher": "^14.3",
+        "laravel-lang/common": "^3.1",
         "laravel/pint": "^1.0",
-        "laravel/sail": "^1.0.1",
+        "laravel/sail": "^1.18",
         "mockery/mockery": "^1.4.4",
-        "nunomaduro/collision": "^6.1",
-        "nunomaduro/larastan": "^2.0",
-        "phpunit/phpunit": "^9.5.10",
-        "spatie/laravel-ignition": "^1.0"
+        "nunomaduro/collision": "^7.0",
+        "nunomaduro/larastan": "^2.6",
+        "phpunit/phpunit": "^10.1",
+        "spatie/laravel-ignition": "^2.0"
     },
     "autoload": {
         "psr-4": {
@@ -77,9 +71,10 @@
         "preferred-install": "dist",
         "sort-packages": true,
         "allow-plugins": {
-            "pestphp/pest-plugin": true
+            "pestphp/pest-plugin": true,
+            "php-http/discovery": true
         }
     },
-    "minimum-stability": "dev",
+    "minimum-stability": "stable",
     "prefer-stable": true
 }
