@@ -3,37 +3,60 @@
 namespace App\Exceptions;
 
 use Exception;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
 use Throwable;
 
 class TranslatableException extends Exception
 {
     /**
-     * Detailed and not user-friendly error message. Targeted to the developer.
+     * User-friendly error message. Targeted to the end-user.
      *
      * @var string
      */
-    protected $error;
+    protected string $friendlyMessage;
+
+    /**
+     * Determine if the exception should be reported.
+     *
+     * @var bool
+     */
+    protected bool $shouldReport;
+
+    /**
+     * Contextual information about the error.
+     *
+     * @var array<int|string, mixed>
+     */
+    protected array $context = [];
 
     /**
      * Create a new exception instance.
      *
-     * @param int            $status
-     * @param string         $error
-     * @param string         $message
-     * @param null|Throwable $previous
-     * @param bool           $isMessageTranslatable
+     * @param int                      $status
+     * @param string                   $error
+     * @param string                   $message
+     * @param null|Throwable           $previous
+     * @param bool                     $messageTranslatable
+     * @param array<int|string, mixed> $context
+     * @param bool                     $shouldReport
      *
      * @return void
      */
-    public function __construct(int $status, string $error, string $message, ?Throwable $previous = null, $isMessageTranslatable = true)
-    {
-        $this->error = $error;
+    public function __construct(
+        int $status,
+        string $error,
+        string $message,
+        ?Throwable $previous = null,
+        bool $messageTranslatable = true,
+        array $context = [],
+        bool $shouldReport = true
+    ) {
+        $this->shouldReport = $shouldReport;
+        $this->friendlyMessage = $messageTranslatable ? strval(__($message)) : $message;
+        $this->context = $context;
 
-        if ($isMessageTranslatable) {
-            $message = strval(__($message));
-        }
-
-        parent::__construct($message, $status, $previous);
+        parent::__construct($error, $status, $previous);
     }
 
     /**
@@ -43,11 +66,37 @@ class TranslatableException extends Exception
      *
      * @return \Illuminate\Http\JsonResponse
      */
-    public function render($request)
+    public function render(Request $request): JsonResponse
     {
         return response()->json([
-            'error'   => $this->error,
-            'message' => $this->getMessage(),
+            'error'   => $this->getMessage(),
+            'message' => $this->friendlyMessage,
         ], $this->getCode());
+    }
+
+    /**
+     * Get the context of the error.
+     *
+     * @return array<int|string, mixed>
+     */
+    public function context(): array
+    {
+        return $this->context;
+    }
+
+    /**
+     * Determine if the exception should use a custom logic for reporting.
+     *
+     * Notice that the method name is misleading, you'd the return value to
+     * determine if the exception should be reported or not. But returning
+     * false means there's no custom logic and the report will use the standard
+     * report method.
+     * More at https://laravel.com/docs/10.x/errors#renderable-exceptions
+     *
+     * @return bool
+     */
+    public function report(): bool
+    {
+        return !$this->shouldReport;
     }
 }
