@@ -37,25 +37,25 @@ class Kernel extends HttpKernel
             \Illuminate\View\Middleware\ShareErrorsFromSession::class,
             \App\Http\Middleware\VerifyCsrfToken::class,
             \Illuminate\Routing\Middleware\SubstituteBindings::class,
-            'localize',
+            \App\Http\Middleware\Localize::class,
         ],
 
         'api' => [
             // \Laravel\Sanctum\Http\Middleware\EnsureFrontendRequestsAreStateful::class,
-            'throttle:api',
+            \Illuminate\Routing\Middleware\ThrottleRequests::class . ':api',
             \Illuminate\Routing\Middleware\SubstituteBindings::class,
-            'localize',
+            \App\Http\Middleware\Localize::class,
         ],
     ];
 
     /**
-     * The application's route middleware.
+     * The application's middleware aliases.
      *
-     * These middleware may be assigned to groups or used individually.
+     * Aliases may be used instead of class names to conveniently assign middleware to routes and groups.
      *
      * @var array<string, class-string|string>
      */
-    protected $routeMiddleware = [
+    protected $middlewareAliases = [
         'auth'                => \App\Http\Middleware\Authenticate::class,
         'auth.basic'          => \Illuminate\Auth\Middleware\AuthenticateWithBasicAuth::class,
         'auth.session'        => \Illuminate\Session\Middleware\AuthenticateSession::class,
@@ -65,9 +65,7 @@ class Kernel extends HttpKernel
         'password.confirm'    => \Illuminate\Auth\Middleware\RequirePassword::class,
         'signed'              => \App\Http\Middleware\ValidateSignature::class,
         'throttle'            => \Illuminate\Routing\Middleware\ThrottleRequests::class,
-        'verified'            => \Illuminate\Auth\Middleware\EnsureEmailIsVerified::class,
-        'localize'            => \App\Http\Middleware\Localize::class,
-        'email_verified'      => \App\Http\Middleware\EmailVerified::class,
+        'verified'            => \App\Http\Middleware\EnsureEmailIsVerified::class,
         'throttle_login'      => \App\Http\Middleware\ThrottleLogin::class,
         'block-in-production' => \App\Http\Middleware\BlockInProduction::class,
     ];
