@@ -14,11 +14,11 @@ return [
     |
     */
 
-    'paths'                    => ['*'],
+    'paths' => ['*'],
 
-    'allowed_methods'          => ['*'],
+    'allowed_methods' => ['*'],
 
-    'allowed_origins'          => [
+    'allowed_origins' => [
         '*',
         // 'null',
         // 'https://my.example.domain',
@@ -31,11 +31,11 @@ return [
         // '/my-example-string/',
     ],
 
-    'allowed_headers'          => ['*'],
+    'allowed_headers' => ['*'],
 
-    'exposed_headers'          => [],
+    'exposed_headers' => [],
 
-    'max_age'                  => 0,
+    'max_age' => 0,
 
-    'supports_credentials'     => false,
+    'supports_credentials' => false,
 ];
