@@ -1,18 +1,9 @@
 <?php
 
-return [
-    /*
-    |--------------------------------------------------------------------------
-    | Authentication Language Lines
-    |--------------------------------------------------------------------------
-    |
-    | The following language lines are used during authentication for various
-    | messages that we need to display to the user. You are free to modify
-    | these language lines according to your application's requirements.
-    |
-    */
+declare(strict_types=1);
 
+return [
     'failed'   => 'These credentials do not match our records.',
-    'password' => 'The provided password is incorrect.',
+    'password' => 'The password is incorrect.',
     'throttle' => 'Too many login attempts. Please try again in :seconds seconds.',
 ];
