@@ -1,98 +1,108 @@
 <?php
 
+declare(strict_types=1);
+
 return [
-    'accepted'             => 'O campo :attribute deve ser aceito.',
-    'accepted_if'          => 'O :attribute deve ser aceito quando :other for :value.',
-    'active_url'           => 'O campo :attribute deve conter uma URL válida.',
-    'after'                => 'O campo :attribute deve conter uma data posterior a :date.',
-    'after_or_equal'       => 'O campo :attribute deve conter uma data superior ou igual a :date.',
-    'alpha'                => 'O campo :attribute deve conter apenas letras.',
-    'alpha_dash'           => 'O campo :attribute deve conter apenas letras, números e traços.',
-    'alpha_num'            => 'O campo :attribute deve conter apenas letras e números .',
-    'array'                => 'O campo :attribute deve conter um array.',
-    'before'               => 'O campo :attribute deve conter uma data anterior a :date.',
-    'before_or_equal'      => 'O campo :attribute deve conter uma data inferior ou igual a :date.',
-    'between'              => [
+    'accepted'        => 'O campo :attribute deve ser aceito.',
+    'accepted_if'     => 'O :attribute deve ser aceito quando :other for :value.',
+    'active_url'      => 'O campo :attribute deve conter uma URL válida.',
+    'after'           => 'O campo :attribute deve conter uma data posterior a :date.',
+    'after_or_equal'  => 'O campo :attribute deve conter uma data superior ou igual a :date.',
+    'alpha'           => 'O campo :attribute deve conter apenas letras.',
+    'alpha_dash'      => 'O campo :attribute deve conter apenas letras, números e traços.',
+    'alpha_num'       => 'O campo :attribute deve conter apenas letras e números .',
+    'array'           => 'O campo :attribute deve conter um array.',
+    'ascii'           => 'O :attribute deve conter apenas caracteres alfanuméricos de byte único e símbolos.',
+    'before'          => 'O campo :attribute deve conter uma data anterior a :date.',
+    'before_or_equal' => 'O campo :attribute deve conter uma data inferior ou igual a :date.',
+    'between'         => [
         'array'   => 'O campo :attribute deve conter de :min a :max itens.',
         'file'    => 'O campo :attribute deve conter um arquivo de :min a :max kilobytes.',
         'numeric' => 'O campo :attribute deve conter um número entre :min e :max.',
         'string'  => 'O campo :attribute deve conter entre :min a :max caracteres.',
     ],
-    'boolean'              => 'O campo :attribute deve conter o valor verdadeiro ou falso.',
-    'confirmed'            => 'A confirmação para o campo :attribute não coincide.',
-    'current_password'     => 'A senha está incorreta.',
-    'date'                 => 'O campo :attribute não contém uma data válida.',
-    'date_equals'          => 'O campo :attribute deve ser uma data igual a :date.',
-    'date_format'          => 'A data informada para o campo :attribute não respeita o formato :format.',
-    'declined'             => 'O :attribute deve ser recusado.',
-    'declined_if'          => 'O :attribute deve ser recusado quando :other for :value.',
-    'different'            => 'Os campos :attribute e :other devem conter valores diferentes.',
-    'digits'               => 'O campo :attribute deve conter :digits dígitos.',
-    'digits_between'       => 'O campo :attribute deve conter entre :min a :max dígitos.',
-    'dimensions'           => 'O valor informado para o campo :attribute não é uma dimensão de imagem válida.',
-    'distinct'             => 'O campo :attribute contém um valor duplicado.',
-    'doesnt_end_with'      => 'O :attribute não pode terminar com um dos seguintes: :values.',
-    'doesnt_start_with'    => 'O :attribute não pode começar com um dos seguintes: :values.',
-    'email'                => 'O campo :attribute não contém um endereço de email válido.',
-    'ends_with'            => 'O campo :attribute deve terminar com um dos seguintes valores: :values',
-    'enum'                 => 'O :attribute selecionado é inválido.',
-    'exists'               => 'O valor selecionado para o campo :attribute é inválido.',
-    'file'                 => 'O campo :attribute deve conter um arquivo.',
-    'filled'               => 'O campo :attribute é obrigatório.',
-    'gt'                   => [
+    'boolean'           => 'O campo :attribute deve conter o valor verdadeiro ou falso.',
+    'confirmed'         => 'A confirmação para o campo :attribute não coincide.',
+    'current_password'  => 'A senha está incorreta.',
+    'date'              => 'O campo :attribute não contém uma data válida.',
+    'date_equals'       => 'O campo :attribute deve ser uma data igual a :date.',
+    'date_format'       => 'A data informada para o campo :attribute não respeita o formato :format.',
+    'decimal'           => 'O :attribute deve ter :decimal casas decimais.',
+    'declined'          => 'O :attribute deve ser recusado.',
+    'declined_if'       => 'O :attribute deve ser recusado quando :other for :value.',
+    'different'         => 'Os campos :attribute e :other devem conter valores diferentes.',
+    'digits'            => 'O campo :attribute deve conter :digits dígitos.',
+    'digits_between'    => 'O campo :attribute deve conter entre :min a :max dígitos.',
+    'dimensions'        => 'O valor informado para o campo :attribute não é uma dimensão de imagem válida.',
+    'distinct'          => 'O campo :attribute contém um valor duplicado.',
+    'doesnt_end_with'   => 'O :attribute não pode terminar com um dos seguintes: :values.',
+    'doesnt_start_with' => 'O :attribute não pode começar com um dos seguintes: :values.',
+    'email'             => 'O campo :attribute não contém um endereço de email válido.',
+    'ends_with'         => 'O campo :attribute deve terminar com um dos seguintes valores: :values',
+    'enum'              => 'O :attribute selecionado é inválido.',
+    'exists'            => 'O valor selecionado para o campo :attribute é inválido.',
+    'file'              => 'O campo :attribute deve conter um arquivo.',
+    'filled'            => 'O campo :attribute é obrigatório.',
+    'gt'                => [
         'array'   => 'O campo :attribute deve ter mais que :value itens.',
         'file'    => 'O arquivo :attribute deve ser maior que :value kilobytes.',
         'numeric' => 'O campo :attribute deve ser maior que :value.',
         'string'  => 'O campo :attribute deve ser maior que :value caracteres.',
     ],
-    'gte'                  => [
+    'gte' => [
         'array'   => 'O campo :attribute deve ter :value itens ou mais.',
         'file'    => 'O arquivo :attribute deve ser maior ou igual a :value kilobytes.',
         'numeric' => 'O campo :attribute deve ser maior ou igual a :value.',
         'string'  => 'O campo :attribute deve ser maior ou igual a :value caracteres.',
     ],
-    'image'                => 'O campo :attribute deve conter uma imagem.',
-    'in'                   => 'O campo :attribute não contém um valor válido.',
-    'in_array'             => 'O campo :attribute não existe em :other.',
-    'integer'              => 'O campo :attribute deve conter um número inteiro.',
-    'ip'                   => 'O campo :attribute deve conter um IP válido.',
-    'ipv4'                 => 'O campo :attribute deve conter um IPv4 válido.',
-    'ipv6'                 => 'O campo :attribute deve conter um IPv6 válido.',
-    'json'                 => 'O campo :attribute deve conter uma string JSON válida.',
-    'lt'                   => [
+    'image'     => 'O campo :attribute deve conter uma imagem.',
+    'in'        => 'O campo :attribute não contém um valor válido.',
+    'in_array'  => 'O campo :attribute não existe em :other.',
+    'integer'   => 'O campo :attribute deve conter um número inteiro.',
+    'ip'        => 'O campo :attribute deve conter um IP válido.',
+    'ipv4'      => 'O campo :attribute deve conter um IPv4 válido.',
+    'ipv6'      => 'O campo :attribute deve conter um IPv6 válido.',
+    'json'      => 'O campo :attribute deve conter uma string JSON válida.',
+    'lowercase' => 'O :attribute deve ser minúsculo.',
+    'lt'        => [
         'array'   => 'O campo :attribute deve ter menos que :value itens.',
         'file'    => 'O arquivo :attribute ser menor que :value kilobytes.',
         'numeric' => 'O campo :attribute deve ser menor que :value.',
         'string'  => 'O campo :attribute deve ser menor que :value caracteres.',
     ],
-    'lte'                  => [
+    'lte' => [
         'array'   => 'O campo :attribute não deve ter mais que :value itens.',
         'file'    => 'O arquivo :attribute ser menor ou igual a :value kilobytes.',
         'numeric' => 'O campo :attribute deve ser menor ou igual a :value.',
         'string'  => 'O campo :attribute deve ser menor ou igual a :value caracteres.',
     ],
-    'mac_address'          => 'O :attribute deve ser um endereço MAC válido.',
-    'max'                  => [
+    'mac_address' => 'O :attribute deve ser um endereço MAC válido.',
+    'max'         => [
         'array'   => 'O campo :attribute deve conter no máximo :max itens.',
         'file'    => 'O campo :attribute não pode conter um arquivo com mais de :max kilobytes.',
         'numeric' => 'O campo :attribute não pode conter um valor superior a :max.',
         'string'  => 'O campo :attribute não pode conter mais de :max caracteres.',
     ],
-    'max_digits'           => 'O :attribute não deve ter mais que :max dígitos.',
-    'mimes'                => 'O campo :attribute deve conter um arquivo do tipo: :values.',
-    'mimetypes'            => 'O campo :attribute deve conter um arquivo do tipo: :values.',
-    'min'                  => [
+    'max_digits' => 'O :attribute não deve ter mais que :max dígitos.',
+    'mimes'      => 'O campo :attribute deve conter um arquivo do tipo: :values.',
+    'mimetypes'  => 'O campo :attribute deve conter um arquivo do tipo: :values.',
+    'min'        => [
         'array'   => 'O campo :attribute deve conter no mínimo :min itens.',
         'file'    => 'O campo :attribute deve conter um arquivo com no mínimo :min kilobytes.',
         'numeric' => 'O campo :attribute deve conter um número superior ou igual a :min.',
         'string'  => 'O campo :attribute deve conter no mínimo :min caracteres.',
     ],
-    'min_digits'           => 'O :attribute deve ter pelo menos :min dígitos.',
-    'multiple_of'          => 'O :attribute deve ser um múltiplo de :value',
-    'not_in'               => 'O campo :attribute contém um valor inválido.',
-    'not_regex'            => 'O formato do valor :attribute é inválido.',
-    'numeric'              => 'O campo :attribute deve conter um valor numérico.',
-    'password'             => [
+    'min_digits'       => 'O :attribute deve ter pelo menos :min dígitos.',
+    'missing'          => 'O campo :attribute deve estar ausente.',
+    'missing_if'       => 'O campo :attribute deve estar ausente quando :other for :value.',
+    'missing_unless'   => 'O campo :attribute deve estar ausente, a menos que :other seja :value.',
+    'missing_with'     => 'O campo :attribute deve estar ausente quando :values estiver presente.',
+    'missing_with_all' => 'O campo :attribute deve estar ausente quando :values estiverem presentes.',
+    'multiple_of'      => 'O :attribute deve ser um múltiplo de :value',
+    'not_in'           => 'O campo :attribute contém um valor inválido.',
+    'not_regex'        => 'O formato do valor :attribute é inválido.',
+    'numeric'          => 'O campo :attribute deve conter um valor numérico.',
+    'password'         => [
         'letters'       => 'O :attribute deve conter pelo menos uma letra.',
         'mixed'         => 'O :attribute deve conter pelo menos uma letra maiúscula e uma minúscula.',
         'numbers'       => 'O :attribute deve conter pelo menos um número.',
@@ -121,14 +131,16 @@ return [
         'numeric' => 'O campo :attribute deve conter o número :size.',
         'string'  => 'O campo :attribute deve conter :size caracteres.',
     ],
-    'starts_with'          => 'O campo :attribute deve começar com um dos seguintes valores: :values',
-    'string'               => 'O campo :attribute deve ser uma string.',
-    'timezone'             => 'O campo :attribute deve conter um fuso horário válido.',
-    'unique'               => 'O valor informado para o campo :attribute já está em uso.',
-    'uploaded'             => 'Falha no Upload do arquivo :attribute.',
-    'url'                  => 'O formato da URL informada para o campo :attribute é inválido.',
-    'uuid'                 => 'O campo :attribute deve ser um UUID válido.',
-    'attributes'           => [
+    'starts_with' => 'O campo :attribute deve começar com um dos seguintes valores: :values',
+    'string'      => 'O campo :attribute deve ser uma string.',
+    'timezone'    => 'O campo :attribute deve conter um fuso horário válido.',
+    'ulid'        => 'O :attribute deve ser um ULID válido.',
+    'unique'      => 'O valor informado para o campo :attribute já está em uso.',
+    'uploaded'    => 'Falha no Upload do arquivo :attribute.',
+    'uppercase'   => 'O :attribute deve ser maiúsculo.',
+    'url'         => 'O formato da URL informada para o campo :attribute é inválido.',
+    'uuid'        => 'O campo :attribute deve ser um UUID válido.',
+    'attributes'  => [
         'address'                  => 'endereço',
         'age'                      => 'idade',
         'amount'                   => 'motante',
