@@ -2,6 +2,7 @@
 
 namespace Tests\Feature\Api\Resources;
 
+use App\Enums\LanguageLineGroup;
 use App\Models\LanguageLine;
 use Tests\TestCase;
 
@@ -10,19 +11,11 @@ use Tests\TestCase;
  *
  * @group Api\Resources
  *
- * @coversNothing
+ * @covers \App\Http\Controllers\Api\ResourcesController::getLanguageLinesByGroup
+ * @covers \App\Http\Resources\Models\LanguageLineResource::toArray
  */
 class GetLanguageLinesByGroupTest extends TestCase
 {
-    use DataProvider;
-
-    /**
-     * The method for the route endpoint.
-     *
-     * @var string
-     */
-    private const Method = 'GET';
-
     /**
      * The route endpoint.
      *
@@ -45,31 +38,22 @@ class GetLanguageLinesByGroupTest extends TestCase
 
     /**
      * Asserts the route returns an error when accesing an invalid group.
-     *
-     * @covers \App\Http\Controllers\Api\ResourcesController::getLanguageLinesByGroup
-     *
-     * @return void
      */
     public function testReturnsErrorOnInvalidGroup(): void
     {
-        $response = $this->json(self::Method, strtr(self::Endpoint, ['{group}' => 'invalid-group']));
+        $response = $this->getJson(strtr(self::Endpoint, ['{group}' => 'invalid-group']));
 
         $response->assertNotFound();
     }
 
     /**
      * Asserts the route returns an empty array when passing an empty group.
-     *
-     * @covers \App\Http\Controllers\Api\ResourcesController::getLanguageLinesByGroup
-     *
-     * @return void
      */
     public function testReturnsEmptyArrayOnEmptyGroup(): void
     {
-        // Assert that no API line exists
-        $this->assertDatabaseMissing('language_lines', ['group' => 'api']);
-
-        $response = $this->json(self::Method, strtr(self::Endpoint, ['{group}' => 'api']));
+        $response = $this->getJson(strtr(self::Endpoint, [
+            '{group}' => LanguageLineGroup::randomCase()->value,
+        ]));
 
         $response
             ->assertOk()
@@ -78,35 +62,25 @@ class GetLanguageLinesByGroupTest extends TestCase
 
     /**
      * Asserts the route returns a valid array for a given group.
-     *
-     * @covers \App\Http\Controllers\Api\ResourcesController::getLanguageLinesByGroup
-     * @covers \App\Http\Resources\Models\LanguageLineResource::toArray
-     *
-     * @return void
      */
     public function testReturnsSuccessOnValidGroup(): void
     {
-        // Create a API language line;
-        LanguageLine::factory()->create([
-            'group' => 'api',
-            'key'   => 'EXEMPLE_KEY',
-            'text'  => [
-                'en' => 'Example Text',
-            ],
-        ]);
+        $locale = config('app.locale');
+        assert(is_string($locale));
 
-        // Assert the API line exists;
-        $this->assertDatabaseCount('language_lines', 1);
-        $this->assertDatabaseHas('language_lines', ['group' => 'api']);
+        $languageLine = LanguageLine::factory()
+            ->withLocale($locale)
+            ->create();
 
-        $response = $this->json(self::Method, strtr(self::Endpoint, ['{group}' => 'api']));
+        $response = $this
+            ->getJson(strtr(self::Endpoint, ['{group}' => $languageLine->group->value]));
 
         $response
             ->assertOk()
             ->assertExactJson([
                 [
-                    'key'  => 'EXEMPLE_KEY',
-                    'text' => 'Example Text',
+                    'key'  => $languageLine->key,
+                    'text' => $languageLine->text[$locale],
                 ],
             ]);
     }
@@ -114,43 +88,32 @@ class GetLanguageLinesByGroupTest extends TestCase
     /**
      * Asserts the route respects the Accept-Language header.
      *
-     * @covers \App\Http\Controllers\Api\ResourcesController::getLanguageLinesByGroup
-     * @covers \App\Http\Resources\Models\LanguageLineResource::toArray
-     *
-     * @dataProvider localizedLanguageLine
+     * @dataProvider App\Enums\Language::asDataProvider
      *
      * @param string $language
-     * @param string $language_line
-     *
-     * @return void
      */
-    public function testRespectsAcceptLanguageHeader(string $language, string $language_line): void
+    public function testRespectsAcceptLanguageHeader(string $language): void
     {
         // Create a API language line;
-        LanguageLine::factory()->create([
-            'group' => 'api',
-            'key'   => 'EXEMPLE_KEY',
-            'text'  => [
-                'en'    => 'Example Text',
-                'pt_BR' => 'Texto de Exemplo',
+        $languageLine = LanguageLine::factory()->create([
+            'text' => [
+                'en'    => fake()->sentence(),
+                'pt_BR' => fake()->sentence(),
             ],
         ]);
 
-        $response = $this->json(
-            self::Method,
-            strtr(self::Endpoint, [
-                '{group}' => 'api',
-            ]),
-            [],
-            ['Accept-Language' => $language]
-        );
+        $response = $this
+            ->withHeader('Accept-Language', $language)
+            ->get(
+                strtr(self::Endpoint, ['{group}' => $languageLine->group->value]),
+            );
 
         $response
             ->assertOk()
             ->assertExactJson([
                 [
-                    'key'  => 'EXEMPLE_KEY',
-                    'text' => $language_line,
+                    'key'  => $languageLine->key,
+                    'text' => $languageLine->text[$language],
                 ],
             ]);
     }
